Return an absolute path for supporting files
When render is cleaning up, it uses this path without modification. This will work if the render is happening in the input directory, but if you render to another directory, the relative path will not resolve and a crufty directory will be created when rendering a self contained format like pdf.

(e.g. `quarto render docs/test-jupyter.md —to pdf`