Support for server-initiated Shiny-to-OJS exporting
This commit makes it possible for OJS code to depend on Shiny variables
that are not known by the client at the time that OJS is starting up.
Normally, if an OJS cell depends on a variable that is not known, you
get an "X is not defined" error; with this change, the OJS runtime
starts out in a lenient phase where undefined variables are treated as
pending promises. Only after the Shiny server function has sent its
first message does the lenient phase end, and undefined variables
become errors again.