windows: when necessary, discover R using registry curent version
Note that R_HOME and any version of R on the PATH are preferred to what is in the registry. Note also that R front ends like RStudio and R GUI will automatially set R_HOME, so when Quarto is run from within these environments it will use the same version of R as the host session.