Avoid using `auto` as min size in grid
When min is set to 'auto', once the size of the grid column wants to reduce from the max, the min will be used. In this case 'auto' is an indefinite width driven by the width of the content, so using it will cause the content to force the grid size.

See
https://css-tricks.com/preventing-a-grid-blowout/