Provide a bootstrap customization layer
It needs to layer in the theme layers (from YAML) just after the built in theme (so it will override any of the values, but not any subsequent user themes). If there is no built in theme specified, just place it at the beginning of the layers so any user layers will override it.