reveal: remove a11y plugin
The a11y plugin's approach to hiding slides broke the Reveal Auto Animate feature. Ideally we'd like a solution that is optimal for accessibility and allows all reveal features to continue working.

We'll soon be making a more comprehensive effort to make Reveal slides fully accessible (incluing ensuring that keybard gestures are available for everything). We'll revisit slide visibility at that time (and perhaps make a PR to reveal if we find an approach that works well and preserves all Reveal behaviors/features).