disable deno v8 optimizing compiler
After updating to Deno 1.16.4 we observed a core dump w/ "Trace/breakpoint trap" after rendering ~ 10 documents in quarto-web. A similar bug was reported for parsing large amounts of YAML here (https://github.com/denoland/deno/issues/12885) and the problem was diagnosed as very likely a V8 optmizing compiler bug (and reported upstream here: https://bugs.chromium.org/p/v8/issues/detail?id=12444).

As with the other reported issue, disabling the optimizing compiler with `--v8-flags=--noopt` prevented the crash. I measured and for a render of a simple document this does create a ~ 20% slowdown. We'll take this slowdown for now so that we can sync to the latest version of Deno and assume that it will be resolved upstream soon.