Improve Extended Figure Processing
We previously always did extended figure processing on figure divs, but in commit 1a3697b961cb57ba0318ba09ccb3dbf1d94c78f5 we introduced a change to permit sometimes processing bare image figures without captions (which we previously didn't process). In this change, we ensnared the figure div processing in the same set of conditions, which broke the processing of figure divs (in a lot of cases).

This change reverts the behavior for figure divs to always on unless explicitly turned off, and introduces a more sophisticated filter for bare figures that will process captionless images if fig attributes are present or if column data is provided.