Properly deal with HTML escaping
Note that we also remove the html raw block from the templates. This is because values like the title or description could contain markdown, so we need to permit that to be rendered. As a result, we do need to be careful with whitespace in the EJS templates, since that can cause lines to be treated as code etc...