Improve grid `page-layout` behavior
```
page-layout: <article | full | custom | none>, defaults to article

Article
Article layout provides a content area with a page based grid layout that provides margins, areas for sidebars, and a reading width optimized body region in which the main content is displayed. The precise size of the document regions will vary slightly depending upon the sidebar (if present) and the presence or absence of margin or complex layout elements.

Full
Full layout uses the article grid system, but will automatically expand the content area to use the sidebar and margin region if no content is placed within those regions. This is useful for layouts that don't need to be constrained to reading width and that will benefit from additional content width.

Custom
Custom layout provides an HTML content container with a reasonable set of margins. This is useful when you'd like to create a page layout within a reasonable inset.

None
None provides no HTML content container. This is useful when you'd like to have full control of the page layout including controlling position to the edge of the viewport (e.g. full bleed).
```