Improve latexmk margin position error detection
We originally we were looking for a message like:

`Package marginnote Info: xpos seems to be \@mn@currxpos  on input line 213.`

which indicates that an element is still unpositioned. But in many cases, there will be a subsequent message that resolves this issue, like:

`Package marginnote Info: xpos seems to be 367.46002pt on input line 213.`

Rather than only check for the first message, we actually need to look for that message, then see if the same line is resolved later in the file.