Move Article Layout Process to Finalizer
This will process layout columns once all the extensions have had a chance to manipulate and post process the HTML. This is useful as it will allow all of those to participate in the page grid system.

Note that we are not processing nodes up past 'main' to the root of the page.