Provide better sidebar defaults when no sidebar is defined
`toc-location: left` can now end up using a left sidebar even when the site hasn't defined one. We need to be sure to provide a sane default behavior for the sidebar in the event that this happens (if there is a toc and no sidebar style, select `floating`).

In addition, we should be setting the sidebar bg even when there is no defined sidebar (meaning we should set the sidebar to be body color).