Don't include text in pipeline key
For example, there can be HTML in the item's text. In that case, the `item.text` will not match the `link.innerText` since the HTML is included in the `item.text` but not the `link.innerText`.  For example _books_ inject HTML to denote the chapter number, which breaks usage of text / innerText.

This exposes us to an issue where the same path in a sidebar with different text will not work properly (the pipeline will replace them both with the same text). I will fix this in a subsequent commit.