Update instructions in README
Thank you for the excellent tools and documentation. For users not familiar with how Git and bash scripts work, it would be useful to give explicit instructions that the configure script should be rerun after each instance of `git pull` or `git pull --rebase`. I've added this explicit instruction and sample bash code in the README. Many thanks for considering.