Supplemental Files Can Filter Formats
When a supplemental file is contributed to a render, it can now includes foramts that go along with it (so that it won’t be forced to render in an unsupported format). For example, when a listing gets added to the render of a revealjs document, it still needs to render as an html document rather than as a revealjs presentation.

https://github.com/quarto-dev/quarto-cli/issues/376