Remove Support for `self-contained-math`
We added the ability for users to specify `self-contained` but to not have math (which is quite large) self contained by setting `self-contained-math: true`. But this adds complexity and likely for no great reason (since in general the idea of self-contained files that don’t render math in a self contained way seems like a non-feature except in edge cases).