Page Categories Find Nearest Listing
When the user clicks a category in the title block of a page, we would previously try to navigate to the referrer page to show the categories (in the event that the referred was a listing that contained this item). Now, we will look up the tree for the nearest listing that contains the item and use that if available (then falling back to old behavior).