Improve Listing Description Default Behavior
We were previously reading the description for listings (if a description or abstract wasn’t provided) by looking at the source file’s markdown. This is rife with issues trying to find the proper content. Instead, allow the pages to render and then actually fetch the first paragraph out of the rendered content (and use that!).