Provide more specific selector
RevealJS slides merge the reveal format extras over the top of htmlformatextras (so reveal can override those). As a result, the html format extras CSS/js appears at the top of the document head rather than near the bottom.

In this case, that meant that a specific text-highlight css selector which worked ok in HTML (because the css was at the bottom of the head, overriding inline default styles) did not work in reveal, since it appeared above the inline default styles.

This fix doesn’t resolve that larger issue, but does dodge this specific issue by providing a more specific selector.