Ensure that cache returns a copy
When reading an extension from the cache, we need to be sure to return a copy of the object. This is because downstream, the object could be mutated (in this case, paths are resolved on the object itself). When the cache wasn’t returning a copy, the cached object would be mutated by the first path resolution, and subsequent path resolutions would fail (since the resolved path was now stored in the cache).