Use proper root directory when resolving external resources
The root directory is used to prevent dependencies from being injected from outside the current directory. We were previously using the input directory, but in the case of project renders, we need to use the project directory instead (e.g. any children of a project can refer to anything within the project itself).