Remove `template-partials` from metadata
It is resolved when the template and partials are all staged, so it no longer needs to be present downstream (for pandoc / lua).

Note that the underlying issue this causes is an informational message in pandoc because the strings in the metadata.yml file are treated as markdown, so this string looks like a set of LaTeX raws on windows.