Code tools shouldn’t muck with background color
Since the code is being highlighted, it is possible that it will only be functional against a specifically colored background (especially dark or light). We should respect the background color if we’re going to highlight the code.

Additionally, since the code div is contained within a popup, remove any border radius so it fills the container edge to edge.