Store extension path as absolute
If we don’t, we can be fooled by the working directory (e.g. the question will remain what the extension path is relative to!). The proper structure is:

```
extension:
   path: <Absolute>
   contributes:
      shortcodes: <relative-to-extension>
      filters: <relative-to-extension>
```

This will allow all paths to be resolve no mattter the working directory, etc…

Additionally, ensure that embedded extensions paths are handled appropriately.