Smart glob handling incorrectly ignoring `!`
We are attempting to detect globs, and ! would normally qualify, but we clip off the ‘!’ and pass the rest as an exclude. As a result, the smart glob handling doesn’t kick in. This change always uses smart globbing for exlcudes unless the called explicitly asks for strict globbing.

Includes unit test.