Invalid href when output-file contains a ‘.’ in its name
If the output the output file is simple, like:

```
output-file: test
```

we automatically would generate the output file `test.html`.

but if the `output-file` contains a `.`, then we would not append the extension (thinking that it already had one), resulting in a case like:

```
output-file: file.name
```

producing a file named `file.name` rather than `file.name.html`.

This will amend the behavior do the appending for any file without an extension or if the appending won’t result in duplicate extensions (e.g. `file.html.html`).