Fix flickering sidebar collapsed contents
When display is set to none, the height effectively becomes 0, so the content may no longer overlay hidden regions (since the collapsed regions height effectively becomes 0). We were doing this to prevent ‘scrollable content’ from showing scrollbars even when the content wasn’t visible, so instead, just toggle the overflow, which should result in the sidebars having height, not being visible, and not showing scrollbars.

Addresses: #2019, #1995