options.output appears to be absolute not relative
We’re treating this path as if it is relative, but it appears that the path created here:

https://github.com/quarto-dev/quarto-cli/blame/main/src/command/render/output.ts#L55

Appears to be always an absolute path. Treating it as an absolute path appears to resolve the issue though I’m a little nervous about side effects given how long this has been in product.