Exclude the project output directory when resolving resources
We are using smart globs to resolve resource paths that are specified in _quarto.yml. A side effect of this is that a simple path like `foo.css` will be ‘smart globbed’ into `**/foo.css` which will then match `foo.css` just about anywhere (including in `_site` or `_book`). We could stop start globbing, but that isn’t that great since it is nice to allow users to write simple directory paths like `my-dir` rather than `my-dir/**`.

Instead, this commit will exclude the project output directory from the list of places that we discover resources.

Fixes #2097