Improve Author Parsing
If the author name was specified as empty string like so:

```
author: “”
```

First we would throw an error since we assumed that all author names would return some value from a bibtex render of the names (into their parts). This isn’t a safe assumption, now we are more careful about checking for nil.

Once error was fixed, we would actually print `author: true` which was bizarre, because rather than provide an empty table for authors, we would provide a table with a literal keys who value was empty. Fix that.