Project Extensions
Project extensions allows an extension to contribute metadata to a project file that is being rendered. It is triggered by referencing the extension name in the `type` field of project, like:

```
project:
  type: my-custom-type
```

or

```
project:
  type: my-org/my-custom-type
```

which will load the matching extension and merge contributed project metadata into the project before files are rendered. This makes it possible to modify project wide settings via an extension.