Improve Handling of TeX Package Installation
We were previously still calling `tlmgr add path` to place binaries in the system path. This is undesirable when we’re running in the embedded tinytex mode. Instead, we will now place the tinytex bin path in the environment when doing PDF compilation so any installed packages or commands will be discoverable.