Refactor extension host resolution
Extension hosts now need to provide a function(regex) to parse a name into an extension host and url resolvers that can create a set of urls for a given host. We’ll use that definition to create a set of resolvers that we can iterate through (and actually fetch their resolved extension url) until we find a resolver that can provide the extension.

- Also make unknown url less dumb
- Also prep for subdirectories