Set scrolltop
clicking on anchor links (links within the same page) causes scrolling of the page. If this causes a scroll upwards, Headroom will pin the header. This can cause the linked element to be obscured by the header.

To avoid this problem we can define  scroll-padding in CSS, which is designed for just this purpose:

Fixes #2641