Process Footnotes in Figure Captions
This will properly shift footnotes that we discover outside of the caption itself.

Pandoc does a nice job of detecting this situation and dealing with in the LaTeX writer here:
https://github.com/jgm/pandoc/commit/411a25306c2d9e64a27bbbd935979079140d9703

Because we've taken over the generation of LaTeX captions, we're evading the writer and consequently Pandoc doesn't solve problem so this post processor more or less replicates that behavior.