Improve Listing Reading
When resolving listing contents, we were previously iterating through each content item and loading items from it. This usually worked fine, but in cases where a user would like to exclude a file from the listing, like:

```
contents:
  - blog/**/*.qmd
  - !blog/index.qmd
```

We would treat each of the globs as independent, therefore never excluding the second file.

We were iterating through items since we allow interspersing metadata and globs, like:

```
contents:
  - blog/**/*.qmd
  - title: foo
    description: bar
```

This change makes it so we merge the globs and metadata items and process them independently, as groups. This causes us to resolve all globs together at once.