Book Single File Resource Copying
Single file books that ingest the resource to produce a single output file (PDF, EPUB, DOCX) now will report that they are single file and this will cause any resources not to be copied to the output directory.

LaTeX produces a single file book that needs to point to resources, so in that case the resources will be copied to the output dir.