Fix indeterminant sorting in listing tables
Table listings were including the hyperlink `a` tag itself in the contents of the title, meaning that the filter would also apply to the contents of the a tag (for example, the HREF). The weirdness was especially pronounced when testing local previews and filtering with numbers, since the href would often contain port numbers, making the behavior appear random.