Improve `highlight-style` resolving
We would previously start by treating the `highlight-style` value as a path and seeing whether it exists, then if it did, treating it like a theme path. By default the `highlight-style` value is `arrow`, so a folder named `arrow` sitting next to a QMD would cause this issue :(.

Instead, first try to resolve it as a built in theme, then search the file system if it doesn’t resolve!

Fixes #3739