Add support for `content-mode` in the grid system
By default, Quarto’s grid system looks and the contents of the sidebar and margin to determine whether to use a standard width content grid, a slim content grid, or a full content grid.

This is an advanced feature that would allow users to specify which content grid to use, overriding Quarto’s default behavior which is `auto`. For example:

```
grid:
  content-mode: standard
```

Will force the use of the standard grid, even when there is margin content present.

This could be useful in cases where most pages have minimal or no margin content, but an occasional page includes short asides. Currently the page layout would vary from page to page, but by specifying a standard content-mode, each page would have the same grid layout, even when margin content was present, which prevent content from ‘jumping’ as the grid shifts from page to page.