Collect `unused` annotations and output an ordered list
Fixes #4953

Previously, we would warn users about ‘unused’ annotation items and not output them. This commit changes the behavior to instead collect those unused annotations and output them in an ordered list below the annotated code (with no warning).

This will allow us to support the case of back to back ordered lists, which appear as a single list in our filter.