Provide input + notebook info in embed placeholder
When rendering single file books, if subfolders are being used then the base path used to resolve notebooks will be resolved against the root input rather than the individual documents. We need to be sure we’re resoving the paths to notebooks using the original input path.

So now, include the input path in the placeholder

- The input path will be used to regenerate the notebook cells as well as to form the jupyter assets that are used to store notebook outputs

Since we’re passing the input through, we can just use that on the other side

cool