Simplify Id suffixing using a post processor
It is very complicated to try to assure that the ids are all properly fixed up, especially for things like references, whose ID is derived from the bibliography. Instead of trying to do this in the AST, use a simple line based post process to find and add suffixes to subnotebook ids and rids.