`quarto.doc.add_supporting()`
This adds supporting for `add_supporting`

The distinction between adding a resource and adding a supporting file basically comes down to whether the file should be `cleaned up` after render / publishing. Resources are left in place, so if you generate a resource, then call `add_resource`, the resource will be _copied_ to the output directory, leaving it in place. Supporting files will be _moved_ to the output directory.