Narrow image auto-detection
fixes #6091

It is reasonable to continue to try to narrow the image auto discovery - users may always specify the image explicitely using the front matter if we are unable to automatically discover the image.

We may need to keep narrowing it futher, but let’s see how this holds up.