Properly resolve dereferenced values
If the root keys were undefined, we would ignore them an use the secondary value, resolving it against the root metadata. For example:

```
{{< meta website.title >}}
```

would actually resolve to the root `title` metadata, since the `website` table isn’t present in the metadata.

This will cause us to stop immediately as soon as a value doesn’t resolve to a table or value.