Exclude project context from qmd -> ipynb intermediary render
Some project types filter the allowable output formats (e.g. books can’t produce `ipynb` output). In this case, however, we are rendering the qmd into a purely intermediary ipynb, which should bypass this configuration and filtering.

This change removes the project from the embed qmd -> ipynb rendering, shielding the render from any project specific configuration or options, which is reasonable since the output ipynb isn’t a participant in the project, it is just an intermediary.

Fixes #6499