Don’t add .5em padding at top of first margin item
Fixes

by default, my margin text sits higher than the body text to its left, so I have to add padding-top to move the margin text down a bit. The screenshot shows after the css is applied. Is this expected? Should the margin text sit higher than the body text to it's left by it default?

```
.column-margin figcaption, .margin-caption, div.aside, aside, .column-margin {
  color: #4C5256;
  font-size: .8rem;
  padding-top: 12px;
}
```