Remove Algolia Search Limit / Warning
By default, we limit the number of search results to 20. This number was chosen not because it was the ideal search experience from a result perspective, but because of a limitation in the way that Algolia search analytics works (accepting 20 item ids when viewing search results).

This change properly splits the view items into buckets of 20, allowing us to raise the default number of search results. In addition, this actually appears to fix the item viewed event, which appears to have been using an incorrect property `objectIDs` instead of

Fixes #7105