[knitr] Correctly write markdown content to rmd intermediate on windows
On Windows `markdown` content parsed by Quarto from .qmd file is passed to R part of code base as a single string using windows EOL (`\r\n`).
Previously, using base::write() was writing an empty line between each original linebreak because it is based on `cat()` and does write linebreak for `\r` and `\n`.  We need the rmd intermediate as close as possible to the input .qmd