Overhaul layout handling
Layout attribute can now carry `flow` or `fill` or be omitted altogether.

We’ll use this explicitly set value to compute the grid size, but if it is omitted we’ll `automatically` compute the grid size by inspecting the children for explicitly set sizes, by considering the type of children, or ultimately by inheriting the parent fill.