Provide stable ids for appendix sections
Fixes #5685

I moved the existing ids up to the section element since when targeting the appendix region with a selector one is very like to want to target the whole section, not merely the contents (excluding the title).

I am leaving heading ids intact as by the time that they arrive at the appendix, the id has been assigned and there is no good way for me to know the difference between an automatically assigned id and a user assigned id.