Fix ‘@‘ mastodon warning by improving about link rendering
This fix actually now runs the href through the markdown pipeline, which will have the side effect of supporting shortcodes in about links.

Note that we rely on a gross mutation of aboutPage.links, sad but true.

Also note that we need to hide `about` from the pandoc metadata, as it will complain about the cite even in the metadata.

Fixes #3493