Correct Filter Extension Resolution When Directory of Same Name is Present
That filter list is expected to point to:

- a named Quarto extension
- a lua file
- an executable file

Because of the last one (executable file), when we resolve filter extensions, we have to be careful to not try to resolve a name that points to a file. We were checking for a path existence before continuing down the extension path and short circuiting because of that directory. The fix will be to check whether the existent path is a file or a directory and doing the right thing.

Fixes #7909