Don’t lose project context when rendering embed qmds
The first fix to #6499 just resolved book embed complaints by suppressing the proiect context for the render. But this can lead to weird behavior such as losing execution context. For example #6520.

This fix preserves the project-ness of the render, but suppressing the project level format enforcement that was causing issues.

Fixes #6520