Improve use/add cmd path resolution
Previously, if a remote extension source provided a target directory (in the form of the repo name), we could always use that when unpacking the repo contents.

Update this to always try that first, but if that path doesn’t exist, proceed and attempt to unpack by looking for a  folder containing an extensions directory or a single lone subdirectory.

This works around an issue (for example) like a repo being renamed. In that case, since the path is redirected, both the old and new repo names will work, but one of the names will not match the folder name in the source tarball. This ensures installation still works in that case.

Fixes #7886