Improve title snipping in notebooks
We promote headings to notebook titles as this is a common pattern in ipynbs. Reducei the footprint where we do this to stop capturing so many headings.

Now we check on the first markdown cell and only do the promotion if there is no title specified and there is no content before the heading.

Fixes #5363
Fixes #6411