Switch to alternative z-index hack for safari
We were previously using translate3d to force safari to deal correctly z-index stacking (by forcing it to recompute stacking). We were relying on a no-op translate to cause this recompute, but that has the side effect of making `fixed` position elements not work, since the position of the root element is being translated. Use of a non-zero opacity has the same effect, but doesn’t affect `fixed` children.

Fixes #7361