Template Partials - Contribute paritals in format extras
When partials are contributed in the format themselves, we can’t tell the different between the user partials and the format contributed partials, so we end up resolving paths on the built in partials.

Instead, contribute in extras, when we know better than to try to resolve the system provided paths.

Fixes #8103