Layout Panels - don’t make cells flex containers
This causes layout issues since each block will be positioned as if it is a part of flexbox layout. Note that `layout-align` really doesn’t work well right now - we need to look more closely at how it should behave and actually make that work. Currently we’re applying a flex style on the individual cell which doesn’t really do anything (it belongs on the parent).

Fixes #8128 which is a regression