Project Render - improve exclude only logic
Rather than use a glob like `**/*` (which will expand to include files in hidden directories or which might otherwise be excluded by engines), or a path like `.` which isn’t a glob and thus doesn’t expand, use the `addDir` function to build the render include list (and then use the excludes to filter this complete list). The `addDir` function will add all the files in the directory, excluding hidden files and engine ignore files, thus doing what we need.

See also:
https://github.com/quarto-dev/quarto-cli/commit/a0e23699ede69996fdf25f4a80ebad23fb182cca

https://github.com/quarto-dev/quarto-cli/commit/196216bebdcbb07828f776c97fc89fce5697eeb1