Improved EJS caching
In Quarto 1.4, we cache the EJS templates for the life of the process. This is usually find, but while developing custom templates using preview, this means that the preview server must be terminated to reload changes in custom templates.

Instead, use the modified time as a simple check to invalidate the cache.  Cleanup the typing while in there.

Fixes #8119