Website Drafts - project options
Add project options to website schema for controlling drafts. Options are:

`visible` - the draft will visible and fully available
`unlinked ` - the draft will be rendered, but will not appear in navigation, search, or listings. (default)
`gone` - the draft will have no content and will not be linked to.

Provides a list of draft documents within the project. This will be used to be injected into website projects via environment or metadata-files and will allow centralized control of whether a file is a draft or not.

Example:

```yaml
project:
  type: website

website:
  title: "Cool Website, Buddy."
  navbar:
    left:
      - stuff/item1.qmd
      - stuff/item2.qmd
      - stuff/item3.qmd
      - listing.qmd
      - text: Another One
  sidebar:
    logo: salute.gif
    align: center
    contents: stuff
    tools:
      - icon: twitter
        href: https://twitter.com
    style: floating
  page-footer:
    center:
      - stuff/item1.qmd
      - stuff/item3.qmd
  draft-mode: unlinked
  drafts:
    - stuff/item3.qmd

format:
  html:
    theme: cosmo
    css: styles.css
    toc: true

```