Project Init - reload context pre-render
When running the pre-render scripts, we have already initialized the project context, which is passed through the rest of the system. The issue with this is that if the pre-render script modifies anything that would appear in the ProjectContext, those changes aren’t reflected downstream.

The simplest example would be generating a file to be included via `metadata-files` - the generated file isn’t present before the pre-render script runs (when the ProjectContext is first loaded), so it doesn’t affect the downstream render as it shows up too late.

This change will load the ProjectContext as normal, but then re-initialize it after the post render script has executed.

Note that certain project data will not be mutable in pre-render, as it is data that is essentially passed to pre-render:

- output-dir
- render list
- projDir
- projType

This is very much included here for review and careful contemplation.