Correctly normalize to forward slash the path when creating tar.gz for Connect deploys
`std/archive/tar.ts` will create folders based on file paths used, but it requires using `/` and not `\` in path. This is a problem on Windows, where the path separator is `\`. This commit adds a check to normalize the path to use `/` instead of `\` when creating the tar.gz file for Connect deploys.