Fix a regression with detecting latex dependencies from R knit_meta in book format
For books, we have a process off creating the includes for dependencies in a later steps, after execution. This steps is done in R but outside of knitting and rendering context, which implies that knitr options should be assumed to be set. All required informations to create the dependencies from knit_meta are passed from TS to R directly.

For other type of render, the dependencies from knit_meta are resolved to includes directly after the knitting process in same R session with same env has R Markdown rendering.

This fix a regression introduced in #7297 while trying to improve the dependencies insertion.