Website Sidebar - global sidebar
We need to wait to automatically assign an id to a sidebar as we use undefined `id` as representing a global sidebar.

Addresses #8869

See also https://github.com/quarto-dev/quarto-cli/issues/5689

We need to wait to automatically assign an id to a sidebar as we use undefined `id` as representing a global sidebar.

Addresses #8869

See also https://github.com/quarto-dev/quarto-cli/issues/5689