Website Listing - file paths
This bug was occurring when we read a file path from YAML. We then see if that file path points to an input file (e.g. a qmd file), in which case we then actually read the QMD file as if it were an item in the listing, using it’s metadata. The `path`, however should be replaced with the resolve path to the input, but we were just keeping the original path from the raw metadata.

We _should_ do this for other fields, so the metadata has the final say, but path should be treated specially.