Website Listings - Ignore Drafts in YAML
When we read YAML files, we actually resolve paths that point to input files (qmd/ipynb) using their metadata. Previously, we would ignore the metadata for draft documents (since they were draft), but leave them in the list, expecting that the user would not include draft documents in yaml listings.

But this isn’t good, since it requires users to coordinate their yaml listings with their draft status.

This change will simply drop any items in YAML which resolve to a document that is a draft.