copy button - initialized clipboardJS for support in modal window
clipboardJS that powers the copy button needs to be aware of the container in focus for it to work. Currently, there is no way to initiallize when clipboardJS instance for handling both button in main body and those in modal windows.

So we need to initialize clipboardJS differently for the modal window.