Faster previewing of .html resource files
Before this commit, the previewing of .html
resource files was slow in larger websites. This
was because the internalPreviewServer was
looking too hard for a .qmd file to render,
including refreshing the project. On the
posit-dev/py-shiny-site repo, doing so is very
slow because expanding the list of resource
files takes multiple seconds.