Set MPL savefig.dpi to figure
Making plots with matplotlib, the dpi of the displayed image (`figure.dpi`)
can be controlled independently of the dpi with which it is saved
(`savefig.dpi`).

If you want the same value for the two cases the recommended way to
do it is to set `figure.dpi` to the desired numeric value, then have
`savefig.dpi` reference the `figure.dpi`. This is a achieved with

   plt.rcParams["savefig.dpi"] = "figure"

Is also the default value in matplotlib.

When `savefig.dpi` is set to the same numeric value as `figure.dpi`, it
may not apply because the matplotlib figure can be saved (and it is often
useful to do so) outside the context in which it was created; in which case
it would refer to another `savefig.dpi` value.

Ref: has2k1/plotnine#773