Respect image width/height set by ipython
First this commit fixes an error where `"key"` is used instead of `key`.

Images created by ipython display functions(which includes matplotlib
plots) may have the desired display width and height in the metadata.
When they do, we fallback to those values instead of none.

After this commit, the display size of images in jupyter notebooks
that are rendered by quarto should be exactly the same as the sizes
in their original environment.

fixes #9470