theme key with format revealjs should never be overriden by user user input
at latest step, we do override some user metadata, and theme is one of theme. However, for revealjs format it has a special handling internally as a quarto theme is built using providing theme key in the metadata. So, we should not override it.

fixes #9548 - revealjs theming broken when `theme` specified at top-level of frontmatter