toc - update active link in toc before expanding
toc expansion is controled by `toc-expand` and computed on client side once at initialization and then at scrolling.  It is based on links in the toc that are active. Default is first link to be the active one.

When no scrolling (body content height smaller than window height),  we need to update the active link to be last section seen on screen, and do this before we expand the toc.