ipynb  - check if svg display data is encoded or not
We assumed svg mime type data was never encoded, but this is not the case when the SVG is produced by a pandoc convert `--to ipynb`. It happens in render context when embed is used. The svg data needs to be decoded before being written to the file at embedding time.