`format-resources` needs to use strict globbing.
This is because we need `format-resources` in extension to be able to move a folder and not just its content.

Without strict mode, any value passed will be transformed to a smartGlob by adding `**/*` to directory to resolve all the file content. Our format-resources process copy the files to root then.

We want 1.3 behavior here where this copy a folder
````yaml
format-resources: folder-from-extensions-to-be-in-root-with-its-content
````