Fix windows quarto.cmd script
Correctly check when QUARTO_DENO_ env var are set and only add flags to it if not present already.

This is necessary for when we call quarto run from Lua inside quarto render as all the environment variables are passed to the child process and so they will be set.

Soùe flags can't be duplicated like `--v8-flags` and create an error when running deno.